+++ CHANGELOG.md
@@ -10,6 +10,8 @@ All notable changes to InventoryDifferent will be documented here.
 
 ### Fixed
 
+- Web and iOS chat: fixed empty collection responses caused by invalid GraphQL field names (storage, operatingSystem, cpu, graphics) that were removed when the data model was restructured into relation arrays
+
 ---
 
 ## [3.0.0] - 2026-07-11